backend/api/views/tags.py
from rest_framework import viewsets, permissions

from core.models import Tag
from api.serializers.tags import TagSerializer


class TagViewSet(viewsets.ReadOnlyModelViewSet):
    queryset = Tag.objects.all()
    serializer_class = TagSerializer
    permission_classes = (permissions.AllowAny,)
    pagination_class = None
